Update migrant
This change fixes a bug where running this file from the /vendor/bin folder wasn't finding the composer autoload file. This was obscured because during testing I was calling this from that folder, but my migrations folder was in <root>/db which worked the same as finding the autoload as if it was a standalone package. This fixes that problem.